% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.cutoff.R
\name{cusum.cutoff}
\alias{cusum.cutoff}
\title{Generates critical values for CUSUM statisitcs.}
\usage{
cusum.cutoff(cusum.obj, upp = 0.975, low = 0.025, Breps = 1000)
}
\arguments{
\item{cusum.obj}{an object returned from cusum or cusum.poly}

\item{upp}{user specified upper tail cut off. Default is .975}

\item{low}{user specified lower tail cut off. Default is .025}

\item{Breps}{number of bootstrap samples}
}
\value{
Returns a matrix of lower and upper cut off values and corresponding standard deviations based on bootstrap sample.
}
\description{
\code{cusum.cutoff} Generates a bootstrap sample for cut-off scores.
}
