% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.flag.R
\name{cusum.flag}
\alias{cusum.flag}
\title{Flags aberrant participants based on CUSUM statistics.}
\usage{
cusum.flag(cusum.obj, cutoff.obj, cut = NULL)
}
\arguments{
\item{cusum.obj}{an object returned from cusum or cusum.poly}

\item{cutoff.obj}{an object returned from cusum.cutoff}

\item{cut}{a vector for user specified cut offs (e.g c(1,1)). The first value is the upper limit. The second value is the lower limit.}
}
\value{
Returns a true or false matrix whether a person is aberrantly responding.
}
\description{
Flags aberrant participants based on CUSUM statistics.
}
