% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.plot.R
\name{cusum.plot}
\alias{cusum.plot}
\title{Generates CUSUM plot for specified IDs.}
\usage{
cusum.plot(cu.object, ID)
}
\arguments{
\item{cu.object}{an object returned from cusum or cusum.poly}

\item{ID}{a numeric ID.}
}
\value{
Returns a plot for specified cusum person chart.
}
\description{
Generates CUSUM plot for specified IDs.
}
