% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex2PL.R
\docType{data}
\name{ex2PL}
\alias{ex2PL}
\title{Example data set based on a simulated 2PL model.}
\format{A data frame with 200 rows and 10 variables.}
\source{
Simulated data.
}
\usage{
data(ex2PL)
}
\description{
Example data set based on a simulated 2PL model.
}
\keyword{datasets}
