% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exGRM.R
\docType{data}
\name{exGRM}
\alias{exGRM}
\title{Example data set based on a simulated GRM model.}
\format{A data frame with 200 rows and 10 variables.}
\source{
Simulated data.
}
\usage{
data(exGRM)
}
\description{
Example data set based on a simulated GRM model.
}
\keyword{datasets}
