% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\docType{data}
\name{gh}
\alias{gh}
\title{Example data set based on a simulated GRM model.}
\format{Gaussian-Hermite Quadature points}
\source{
ltm
}
\usage{
gh
}
\description{
Example data set based on a simulated GRM model.
}
\keyword{datasets}
